+++ packages/registry/app/components/client-components.tsx
@@ -1,20 +1,21 @@
 "use client";
 
-import { paseo, useTx } from "typink";
+import { useTx } from "typink";
 import { WalletSelect } from "@/registry/polkadot-ui/blocks/connect-wallet/components/wallet-select.dedot";
 import { WalletSelect as WalletSelectPapi } from "@/registry/polkadot-ui/blocks/connect-wallet/components/wallet-select.papi";
 import { TxButton as TxButtonDedot } from "@/registry/polkadot-ui/blocks/tx-button/components/tx-button.dedot";
 import { PolkadotProvider as PolkadotProviderDedot } from "@/registry/polkadot-ui/providers/polkadot-provider.dedot";
 
-import { TxButtonStandalone as TxButtonPapi } from "@/registry/polkadot-ui/blocks/tx-button/components/tx-button.standalone.papi";
+import { TxButtonStandalone as TxButtonPapi } from "@/registry/polkadot-ui/blocks/tx-button/components/tx-button.papi";
 import {
   PolkadotProvider as PolkadotProviderPapi,
   usePapi,
-} from "@/registry/polkadot-ui/providers/polkadot-provider.reactive-dot";
+} from "@/registry/polkadot-ui/providers/polkadot-provider.papi";
 import { Binary } from "polkadot-api";
 import { Suspense } from "react";
 import { config } from "@/registry/polkadot-ui/reactive-dot.config";
 import { TxButtonSkeleton } from "@/registry/polkadot-ui/blocks/tx-button/components/tx-button.base";
+import { ClientOnly } from "@/registry/polkadot-ui/blocks/client-only";
 
 export function Components() {
   const tx = useTx((tx) => tx.system.remark, {
@@ -47,36 +48,45 @@ export function Dedot() {
   );
 }
 
-export function Papi() {
-  const networkId = "paseo" as const;
-  const { client } = usePapi();
+export function PapiComponents() {
+  const networkId = "paseo";
+  const { client } = usePapi(networkId);
   const typedApi = client?.getTypedApi(config.chains[networkId].descriptor);
   const transaction = typedApi?.tx.System.remark({
     remark: Binary.fromText("Hello from polkadot-ui!"),
   });
+  return (
+    <>
+      {transaction ? (
+        <ClientOnly>
+          <div className="flex flex-col gap-4 max-w-sm">
+            <WalletSelectPapi />
+            <Suspense
+              fallback={<TxButtonSkeleton>Send Remark (PAPI)</TxButtonSkeleton>}
+            >
+              <TxButtonPapi
+                transaction={transaction}
+                networkId={networkId}
+                notifications={{
+                  title: "Remark Welcome Message",
+                  description: "Please sign the transaction in your wallet",
+                }}
+              >
+                Send Remark (PAPI)
+              </TxButtonPapi>
+            </Suspense>
+          </div>
+        </ClientOnly>
+      ) : null}
+    </>
+  );
+}
 
+export function Papi() {
   return (
     <PolkadotProviderPapi>
       <h1>PAPI</h1>
-      {transaction ? (
-        <div className="flex flex-col gap-4 max-w-sm">
-          <WalletSelectPapi />
-          <Suspense
-            fallback={<TxButtonSkeleton>Send Remark (PAPI)</TxButtonSkeleton>}
-          >
-            <TxButtonPapi
-              transaction={transaction}
-              networkId={networkId}
-              notifications={{
-                title: "Remark Welcome Message",
-                description: "Please sign the transaction in your wallet",
-              }}
-            >
-              Send Remark (PAPI)
-            </TxButtonPapi>
-          </Suspense>
-        </div>
-      ) : null}
+      <PapiComponents />
     </PolkadotProviderPapi>
   );
 }
