+++ packages/registry/components/providers.tsx
@@ -1,12 +1,7 @@
 "use client";
 
-import { PolkadotProvider as PapiProvider } from "@/registry/polkadot-ui/providers/polkadot-provider.papi";
-import { PolkadotProvider as ReactiveProvider } from "@/registry/polkadot-ui/providers/polkadot-provider.reactive-dot";
+import { PolkadotProvider as ReactiveDotProvider } from "@/registry/polkadot-ui/providers/polkadot-provider.papi";
 
 export function Providers({ children }: { children: React.ReactNode }) {
-  return (
-    <ReactiveProvider>
-      <PapiProvider>{children}</PapiProvider>
-    </ReactiveProvider>
-  );
+  return <ReactiveDotProvider>{children}</ReactiveDotProvider>;
 }
