+++ packages/registry/registry/polkadot-ui/blocks/address-input/demo/page.papi.tsx
@@ -1,4 +1,4 @@
-import { PolkadotProvider } from "@/registry/polkadot-ui/providers/polkadot-provider.papi";
+import { PolkadotProvider } from "@/registry/polkadot-ui/providers/__polkadot-provider.papi";
 import { AddressInput } from "@/registry/polkadot-ui/blocks/address-input/components/address-input.papi";
 import { AddressInputDemo } from "@/registry/polkadot-ui/blocks/address-input/demo/address-input-demo";
