+++ packages/registry/registry/polkadot-ui/blocks/connect-wallet/demo/page.papi.tsx
@@ -1,4 +1,4 @@
-import { PolkadotProvider } from "@/registry/polkadot-ui/providers/polkadot-provider.papi";
+import { PolkadotProvider } from "@/registry/polkadot-ui/providers/__polkadot-provider.papi";
 import { ConnectWalletDemo } from "@/registry/polkadot-ui/blocks/connect-wallet/demo/connect-wallet-demo";
 import { WalletSelect } from "@/registry/polkadot-ui/blocks/connect-wallet/components/wallet-select.papi";
