+++ packages/registry/registry/polkadot-ui/blocks/tx-button/components/tx-button.papi.tsx
@@ -1,257 +1,253 @@
 "use client";
 
-import { PolkadotProvider } from "@/registry/polkadot-ui/providers/polkadot-provider.papi";
-import { TxButtonBase, type TxButtonBaseProps } from "./tx-button.base";
-import { usePapi } from "@/registry/polkadot-ui/providers/polkadot-provider.reactive-dot";
-import { useEffect, useMemo, useState } from "react";
+import { Button } from "@/components/ui/button";
+import { cn } from "@/lib/utils";
 import {
-  useChainIds,
-  useClient,
-  useSpendableBalance,
-} from "@reactive-dot/react";
-import { config } from "@/registry/polkadot-ui/reactive-dot.config";
-import type { TxResultLike } from "@/registry/polkadot-ui/lib/types.dot-ui";
-import type { TypedApi } from "polkadot-api";
-import type { SupportedChainId } from "@/registry/polkadot-ui/providers/polkadot-provider.reactive-dot";
+  beginTxStatusNotification,
+  cancelTxStatusNotification,
+  defaultDescriptions,
+  defaultTitles,
+  txStatusNotification,
+} from "@/registry/polkadot-ui/blocks/tx-notification/components/tx-notification.base";
+import { formatBalance } from "@/registry/polkadot-ui/lib/utils.dot-ui";
+import { config as reactiveDotConfig } from "@/registry/polkadot-ui/reactive-dot.config";
+import { useSpendableBalance } from "@reactive-dot/react";
+import { Ban, CheckCheck, CheckCircle, Coins, Loader2 } from "lucide-react";
+import type { Transaction as PapiTransaction } from "polkadot-api";
+import { useEffect, useState } from "react";
+import { TxButtonBaseProps } from "@/registry/polkadot-ui/blocks/tx-button/components/tx-button.base";
+import { ChainId } from "@reactive-dot/core";
+import { usePapi } from "@/registry/polkadot-ui/providers/polkadot-provider.papi";
 
-// Local minimal typings for PAPI transactions built from typed API
-type PapiTx = TypedApi<
-  (typeof config)["chains"][SupportedChainId]["descriptor"]
->["tx"];
+export function TxButtonStandalone(
+  props: TxButtonBaseProps & {
+    transaction: PapiTransaction<object, string, string, unknown>;
+    networkId: ChainId;
+  }
+) {
+  const {
+    transaction,
+    networkId,
+    children,
+    disabled,
+    icons = {
+      default: null,
+      loading: <Loader2 className="w-4 h-4 animate-spin" />,
+      finalized: <CheckCheck className="w-4 h-4" />,
+      inBestBlock: <CheckCircle className="w-4 h-4" />,
+      error: <Ban className="w-4 h-4" />,
+    },
+    notifications = {
+      titles: defaultTitles,
+      descriptions: defaultDescriptions,
+    },
+    ...rest
+  } = props;
 
-interface PapiTxEventBase {
-  type: string;
-  txHash?: string;
-  found?: boolean;
-  isValid?: boolean;
-}
-
-interface PapiTransaction {
-  getEstimatedFees: (from: string) => Promise<bigint>;
-  signSubmitAndWatch: (signer: unknown) => {
-    subscribe: (handlers: {
-      next: (ev: PapiTxEventBase) => void;
-      error: (err: unknown) => void;
-    }) => { unsubscribe: () => void };
-  };
-}
-
-type PapiTxFn = (...args: unknown[]) => PapiTransaction;
-
-// Props type - removes services/tx/execute/args since we inject execute + services
-export interface PapiTxButtonProps
-  extends Omit<TxButtonBaseProps, "services" | "tx" | "args" | "execute"> {
-  txBuilder: (tx: PapiTx) => PapiTxFn;
-  args?: unknown[];
-  networkId?: SupportedChainId;
-}
-
-export function TxButton(props: PapiTxButtonProps) {
-  const { txBuilder, args, networkId, ...rest } = props;
   const { selectedAccount } = usePapi();
+  const signer = selectedAccount?.polkadotSigner;
 
-  const chainIds = useChainIds();
-  const targetChainId: SupportedChainId | undefined = useMemo(() => {
-    if (networkId) return networkId as SupportedChainId;
-    return (chainIds && chainIds[0]) as SupportedChainId | undefined;
-  }, [networkId, chainIds]);
+  const connectedAccount = selectedAccount;
+  const isValidNetwork = true;
 
-  const client = useClient(
-    targetChainId ? { chainId: targetChainId } : undefined
-  );
-
-  const typedApi = useMemo(() => {
-    if (!client || !targetChainId) return null;
-    const descriptor = config.chains[targetChainId]
-      ?.descriptor as (typeof config)["chains"][SupportedChainId]["descriptor"];
-    return client.getTypedApi(descriptor);
-  }, [client, targetChainId]);
-
-  // Fee estimation state
+  type TxStatusLike = { type: string } | null;
   const [fee, setFee] = useState<bigint | null>(null);
   const [isFeeLoading, setIsFeeLoading] = useState<boolean>(false);
   const [feeError, setFeeError] = useState<string | null>(null);
+  const [txStatus, setTxStatus] = useState<TxStatusLike>(null);
+  const [showResult, setShowResult] = useState(false);
+  const [submitError, setSubmitError] = useState<string | null>(null);
+  const balanceFree = useSpendableBalance(selectedAccount?.address ?? "");
 
-  // Build call function lazily
-  const callFn = useMemo(() => {
-    if (!typedApi) return null as unknown as PapiTxFn | null;
-    return txBuilder(typedApi.tx as PapiTx) as unknown as PapiTxFn;
-  }, [typedApi, txBuilder]);
+  const isError = !!submitError || !!feeError || !isValidNetwork;
+  const status = txStatus?.type;
+  const isLoading =
+    status === "Loading" || status === "signed" || status === "broadcasted";
+  const isInBlock = status === "txBestBlocksState" || status === "inBlock";
+  const isFinalized = status === "finalized" || status === "Finalized";
+  const explorerUrl = reactiveDotConfig.chains[networkId].explorerUrl;
+  const symbol = reactiveDotConfig.chains[networkId].symbol;
+  const decimals = reactiveDotConfig.chains[networkId].decimals;
+  const canCoverFee =
+    fee !== null && balanceFree !== null && balanceFree.planck >= fee;
 
-  // Estimate fee whenever inputs change
   useEffect(() => {
-    let cancelled = false;
-    async function run() {
-      setFeeError(null);
-      setFee(null);
-      setIsFeeLoading(true);
-      try {
-        if (!typedApi || !callFn || !selectedAccount?.address) return;
-        const callArgs = (args ?? []) as unknown[];
-        const tx = (callFn as (...fnArgs: unknown[]) => PapiTransaction)(
-          ...callArgs
-        );
-        const estimated = await tx.getEstimatedFees(selectedAccount.address);
-        if (!cancelled) setFee(estimated);
-      } catch (e) {
-        if (!cancelled) setFeeError(e instanceof Error ? e.message : String(e));
-      } finally {
-        if (!cancelled) setIsFeeLoading(false);
-      }
-    }
-    run();
-    return () => {
-      cancelled = true;
-    };
-  }, [typedApi, callFn, selectedAccount?.address, args]);
-
-  // Balance via reactive-dot for can-pay check
-  const balance = useSpendableBalance(
-    (selectedAccount?.address ?? "") as unknown as string,
-    {
-      chainId: targetChainId,
-    } as unknown as { chainId?: SupportedChainId }
-  );
-
-  const nativeToken = useMemo((): { decimals: number; symbol: string } => {
-    if (!targetChainId) return { decimals: 10, symbol: "UNIT" };
-    const token = (
-      config as unknown as {
-        chains: Record<
-          string,
-          { nativeToken?: { decimals?: number; symbol?: string } }
-        >;
-      }
-    ).chains[targetChainId as string]?.nativeToken;
-    return {
-      decimals: token?.decimals ?? 10,
-      symbol: token?.symbol ?? "UNIT",
-    };
-  }, [targetChainId]);
-
-  // Execute function consumed by base
-  const execute = useMemo(() => {
-    return async ({
-      args: execArgs,
-      onStatus,
-    }: {
-      args: unknown[];
-      onStatus: (r: TxResultLike) => void;
-    }) => {
-      if (!typedApi || !callFn) throw new Error("API not ready");
-      if (!selectedAccount?.polkadotSigner)
-        throw new Error("No signer available");
-
-      const tx = (callFn as (...fnArgs: unknown[]) => PapiTransaction)(
-        ...((execArgs ?? []) as unknown[])
+    if (txStatus || isError) {
+      setShowResult(true);
+      const timer = setTimeout(
+        () => setShowResult(false),
+        props.resultDisplayDuration
       );
-
-      await new Promise<void>((resolve, reject) => {
-        try {
-          const sub = tx
-            .signSubmitAndWatch(selectedAccount.polkadotSigner)
-            .subscribe({
-              next: (ev: PapiTxEventBase) => {
-                if (ev.type === "signed")
-                  onStatus({ status: { type: "Signed" } });
-                if (ev.type === "broadcasted")
-                  onStatus({ status: { type: "Broadcasted" } });
-                if (ev.type === "txBestBlocksState") {
-                  if (ev.found) {
-                    onStatus({
-                      status: { type: "InBestBlock" },
-                      txHash: ev.txHash,
-                    });
-                  } else {
-                    onStatus({
-                      status: { type: ev.isValid ? "Valid" : "Invalid" },
-                    });
-                  }
-                }
-                if (ev.type === "finalized") {
-                  onStatus({
-                    status: { type: "Finalized" },
-                    txHash: ev.txHash,
-                  });
-                  sub.unsubscribe();
-                  resolve();
-                }
-              },
-              error: (err: unknown) => reject(err),
-            });
-        } catch (e) {
-          reject(e);
-        }
-      });
-    };
-  }, [callFn, typedApi, selectedAccount?.polkadotSigner]);
-
-  function getPlanckMaybe(value: unknown): bigint | null {
-    if (
-      value &&
-      typeof value === "object" &&
-      "planck" in (value as Record<string, unknown>)
-    ) {
-      const p = (value as Record<string, unknown>).planck as unknown;
-      if (typeof p === "bigint") return p as bigint;
+      return () => clearTimeout(timer);
     }
-    return null;
-  }
+    setShowResult(false);
+  }, [txStatus, isError, props.resultDisplayDuration]);
 
-  const services = useMemo(
-    () => ({
-      connectedAccount: selectedAccount,
-      supportedNetworks: (chainIds ?? []).map((id) => ({
-        id,
-        decimals: ((
-          config as unknown as {
-            chains: Record<string, { nativeToken?: { decimals?: number } }>;
-          }
-        ).chains[id as string]?.nativeToken?.decimals ?? 10) as number,
-        symbol: ((
-          config as unknown as {
-            chains: Record<string, { nativeToken?: { symbol?: string } }>;
-          }
-        ).chains[id as string]?.nativeToken?.symbol ?? "UNIT") as string,
-        name: id,
-      })),
-      balanceFree: getPlanckMaybe(balance as unknown),
-      isFeeLoading,
-      fee,
-      feeError,
-      decimals: nativeToken.decimals,
-      symbol: nativeToken.symbol,
-    }),
-    [
-      selectedAccount,
-      chainIds,
-      balance,
-      isFeeLoading,
-      fee,
-      feeError,
-      nativeToken.decimals,
-      nativeToken.symbol,
-    ]
-  );
+  useEffect(() => {
+    setIsFeeLoading(true);
+    setFeeError(null);
+    setFee(null);
+
+    transaction
+      ?.getEstimatedFees(selectedAccount?.address ?? "")
+      .then((fee) => setFee(fee))
+      .catch((error) =>
+        setFeeError(error instanceof Error ? error.message : String(error))
+      )
+      .finally(() => {
+        setIsFeeLoading(false);
+      });
+  }, [transaction, selectedAccount]);
+
+  if (!signer) return "no signer";
+
+  const handleClick = () => {
+    setSubmitError(null);
+    setTxStatus({ type: "Loading" });
+    console.log("aaa", explorerUrl);
+
+    const toastId = beginTxStatusNotification({
+      network: {
+        name: networkId,
+      },
+      title:
+        notifications.title ??
+        notifications.titles?.signing ??
+        "Waiting for signature...",
+      description:
+        notifications.description ??
+        notifications.descriptions?.signing ??
+        "Please sign the transaction in your wallet",
+    });
+    transaction.signSubmitAndWatch(signer).subscribe({
+      next: (ev) => {
+        console.log(ev);
+        setTxStatus({ type: ev.type });
+        txStatusNotification({
+          title:
+            notifications.title ??
+            notifications.titles?.submitting ??
+            "Waiting for confirmation...",
+          successDuration: props.resultDisplayDuration ?? 10000,
+          result: { txHash: ev.txHash, status: { type: ev.type } },
+          toastId,
+
+          network: {
+            name: networkId,
+            subscanUrl: explorerUrl,
+          },
+        });
+      },
+      error: (err) => {
+        setTxStatus({ type: "Error" });
+        txStatusNotification({
+          title:
+            notifications.title ??
+            notifications.titles?.error ??
+            "Transaction failed",
+          successDuration: props.resultDisplayDuration ?? 10000,
+          result: {
+            status: { type: "Error" },
+          },
+          toastId,
+          network: { name: networkId },
+        });
+        cancelTxStatusNotification({
+          toastId,
+          network: {
+            name: networkId,
+            subscanUrl: explorerUrl,
+          },
+          title:
+            notifications.title ??
+            notifications.titles?.error ??
+            "Transaction failed",
+          description: err instanceof Error ? err.message : String(err),
+        });
+      },
+    });
+  };
+
+  const isButtonDisabled =
+    disabled ||
+    isLoading ||
+    !connectedAccount?.address ||
+    !!feeError ||
+    !isValidNetwork ||
+    !canCoverFee;
 
   return (
-    <TxButtonBase
-      {...(rest as unknown as TxButtonBaseProps)}
-      execute={execute}
-      args={(args ?? []) as unknown[]}
-      networkId={targetChainId}
-      services={services}
-    />
+    <div className="inline-flex flex-col gap-1">
+      <div className="text-xs text-muted-foreground font-normal h-4 flex items-center justify-start">
+        {fee !== null ? (
+          <span className="flex items-center gap-1">
+            <Coins className="w-3 h-3" />
+            {formatBalance({
+              value: fee,
+              decimals: decimals,
+              unit: symbol,
+              nDecimals: 4,
+            })}
+          </span>
+        ) : isFeeLoading ? (
+          <div className="flex items-center">
+            <Loader2 className="w-3 h-3 animate-spin" />
+            <span>Estimating fees...</span>
+          </div>
+        ) : (
+          <span className="text-muted-foreground">
+            Fee calculation pending... {fee} {feeError} {isFeeLoading}
+          </span>
+        )}
+      </div>
+      <Button
+        onClick={handleClick}
+        variant={props.variant}
+        size={props.size}
+        disabled={isButtonDisabled}
+        className={cn(isLoading && "cursor-not-allowed", props.className)}
+        {...rest}
+      >
+        {isLoading ? (
+          <>
+            {children}
+            {icons.loading}
+          </>
+        ) : isInBlock ? (
+          <>
+            {children}
+            {icons.inBestBlock}
+          </>
+        ) : showResult && isFinalized ? (
+          <>
+            {children}
+            {icons.finalized}
+          </>
+        ) : isError && showResult ? (
+          <>
+            {children}
+            {icons.error}
+          </>
+        ) : (
+          <>
+            {children}
+            {icons.default}
+          </>
+        )}
+      </Button>
+      <div className="text-xs font-normal h-4 flex items-center">
+        {!connectedAccount?.address ? (
+          <span className="text-amber-500">Please select an account</span>
+        ) : !isValidNetwork ? (
+          <span className="text-destructive">Invalid network</span>
+        ) : feeError ? (
+          <span className="text-destructive">{feeError}</span>
+        ) : fee !== null && !canCoverFee ? (
+          <span className="text-destructive">Insufficient balance for fee</span>
+        ) : (
+          <span className="text-transparent">_</span>
+        )}
+      </div>
+    </div>
   );
 }
-
-// Wrapped version with provider for drop-in usage
-export function TxButtonWithProvider(props: PapiTxButtonProps) {
-  return (
-    <PolkadotProvider>
-      <TxButton {...props} />
-    </PolkadotProvider>
-  );
-}
-
-TxButtonWithProvider.displayName = "TxButtonWithProvider";
